try another fix to avoid vercel limits for deployment
