+++ api/index.js
@@ -3,7 +3,7 @@ import bodyParser from "body-parser";
 import { connect } from "mongoose";
 import dotenv from "dotenv";
 import { router as categoriesRouter } from "./routes/category.js";
-import { router as userRouter } from "./routes/user.js";
+import { router as userRouter } from "../utils.js";
 import { router as subcategoryRouter } from "./routes/subcategory.js";
 import { router as subcategoryItemsRouter } from "./routes/subcategory-item.js";
