+++ vercel.json
@@ -1,4 +1,4 @@
 {
   "version": 2,
-  "rewrites": [{ "source": "/index.js", "destination": "/api" }]
+  "rewrites": [{ "source": "/(.*)", "destination": "/api" }]
 }
